predict.py
import os
import subprocess
import sys
from typing import List
import torch
from PIL import Image
import numpy as np
import cv2
from cog import BasePredictor, Input, Path
import requests
import zipfile
import io


class Predictor(BasePredictor):
    def setup(self) -> None:
        """Load the model into memory to make running multiple predictions efficient"""
        
        print("🚀 Setting up Stable-Makeup model...")
        
        # Clone the Stable-Makeup repository
        if not os.path.exists("Stable-Makeup"):
            print("📥 Cloning Stable-Makeup repository...")
            subprocess.run([
                "git", "clone", 
                "https://github.com/Xiaojiu-z/Stable-Makeup.git"
            ], check=True)
        
        # Change to the Stable-Makeup directory
        os.chdir("Stable-Makeup")
        
        # Add the current directory to Python path
        sys.path.append(os.getcwd())
        
        # Create models directory
        models_dir = "models/stablemakeup"
        os.makedirs(models_dir, exist_ok=True)
        
        # Download model weights if they don't exist
        self.download_model_weights(models_dir)
        
        # Import and initialize the model
        try:
            from models.stablemakeup import StableMakeupModel
            self.model = StableMakeupModel()
            self.model.load_weights(models_dir)
            print("✅ Model loaded successfully!")
        except ImportError as e:
            print(f"⚠️ Could not import StableMakeup model: {e}")
            print("📝 Using placeholder implementation for now...")
            self.model = None
        
    def download_model_weights(self, models_dir: str):
        """Download the pre-trained model weights"""
        
        # Check if weights already exist
        if os.path.exists(os.path.join(models_dir, "pytorch_model.bin")):
            print("✅ Model weights already exist")
            return
        
        print("📥 Downloading model weights...")
        
        # TODO: Replace with actual Google Drive download link
        # For now, we'll create placeholder files
        placeholder_files = [
            "pytorch_model.bin",
            "pytorch_model_1.bin", 
            "pytorch_model_2.bin",
            "config.json"
        ]
        
        for filename in placeholder_files:
            filepath = os.path.join(models_dir, filename)
            if not os.path.exists(filepath):
                # Create a small placeholder file
                with open(filepath, 'w') as f:
                    f.write(f"# Placeholder for {filename}\n")
                print(f"📝 Created placeholder: {filename}")
        
        print("⚠️ Using placeholder weights. Please download actual weights from:")
        print("🔗 https://github.com/Xiaojiu-z/Stable-Makeup")
        
    def predict(
        self,
        source_image: Path = Input(description="Source face image"),
        reference_image: Path = Input(description="Reference makeup image"),
        makeup_intensity: float = Input(
            description="Makeup transfer intensity (0.1-2.0)",
            default=1.0,
            ge=0.1,
            le=2.0,
        ),
    ) -> Path:
        """Run a single prediction on the model"""
        
        print(f"🎨 Starting makeup transfer with intensity: {makeup_intensity}")
        
        # Load images
        source_img = Image.open(source_image).convert("RGB")
        reference_img = Image.open(reference_image).convert("RGB")
        
        # Resize images to 512x512 (standard for diffusion models)
        source_img = source_img.resize((512, 512))
        reference_img = reference_img.resize((512, 512))
        
        # Save temporary input files
        source_path = "temp_source.jpg"
        reference_path = "temp_reference.jpg"
        output_path = "temp_output.jpg"
        
        source_img.save(source_path)
        reference_img.save(reference_path)
        
        # Run the Stable-Makeup inference
        try:
            if self.model:
                # Use the actual model
                result_image = self.model.transfer_makeup(
                    source_path, 
                    reference_path, 
                    makeup_intensity
                )
            else:
                # Use placeholder implementation
                result_image = self.placeholder_makeup_transfer(
                    source_path, 
                    reference_path, 
                    makeup_intensity
                )
            
            # Save the result
            result_image.save(output_path)
            print("✅ Makeup transfer completed successfully!")
            
            return Path(output_path)
            
        except Exception as e:
            print(f"❌ Error during inference: {e}")
            # Return the source image as fallback
            source_img.save(output_path)
            return Path(output_path)
    
    def placeholder_makeup_transfer(self, source_path: str, reference_path: str, intensity: float) -> Image.Image:
        """
        Placeholder implementation for makeup transfer
        This simulates the effect until the actual model is implemented
        """
        
        # Load images
        source = cv2.imread(source_path)
        reference = cv2.imread(reference_path)
        
        # Convert to RGB
        source_rgb = cv2.cvtColor(source, cv2.COLOR_BGR2RGB)
        reference_rgb = cv2.cvtColor(reference, cv2.COLOR_BGR2RGB)
        
        # Simple color transfer simulation
        # Extract average color from reference image
        ref_mean = np.mean(reference_rgb, axis=(0, 1))
        src_mean = np.mean(source_rgb, axis=(0, 1))
        
        # Apply color adjustment based on intensity
        adjusted = source_rgb.astype(np.float32)
        for i in range(3):  # RGB channels
            adjusted[:, :, i] += (ref_mean[i] - src_mean[i]) * intensity * 0.3
        
        # Clip values and convert back
        adjusted = np.clip(adjusted, 0, 255).astype(np.uint8)
        
        # Convert back to PIL Image
        result = Image.fromarray(adjusted)
        
        return result 