requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.33.2
accelerate==0.23.0
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
scikit-image==0.21.0
face-alignment==1.3.5
mediapipe==0.10.8
facexlib==0.3.0
spiga==0.0.6
dlib==19.24.2
requests==2.31.0
matplotlib==3.7.2
# Cache buster: 2025-07-27-10-52-39
